requirements.txt
alabaster==0.7.16
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
ase==3.23.0
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bibtexparser==1.4.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Deprecated==1.2.14
dnspython==2.6.1
docutils==0.20.1
elastic==5.2.3.6
elastic-transport==8.13.1
elasticsearch==8.14.0
email_validator==2.2.0
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.4
fonttools==4.53.1
free_proxy==1.1.1
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
h3==3.7.7
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
imageio==2.35.0
imageio-ffmpeg==0.5.1
imagesize==1.4.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
Levenshtein==0.25.1
lib3==0.1.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
microsearch==1.0.0
nltk==3.8.1
nominatim==0.1
numpy==1.26.4
orjson==3.10.6
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyarrow==17.0.0
pycparser==2.21
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.18.0
pylance==0.15.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-vlc==3.0.20123
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.4
regex==2023.12.25
requests==2.29.0
rich==13.7.1
scholarly==1.7.11
scipy==1.14.0
selenium==4.23.0
setuptools==69.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
spglib==2.5.0
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
starlette==0.37.2
timezonefinder==6.5.0
tkintervideo==0.0.2
tkVideo==0.1
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.10.0
urllib3==1.26.18
uvicorn==0.30.1
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0